Move CommandPreset to model
