+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -13,7 +13,7 @@ import seedu.address.commons.core.Config;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.events.ui.ExitAppRequestEvent;
 import seedu.address.logic.Logic;
-import seedu.address.logic.commands.CommandPreset;
+import seedu.address.model.CommandPreset;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.task.ReadOnlyTask;
